--- paddle/pten/kernels/cpu/empty_kernel.cc
@@ -1,41 +1,0 @@
-/* Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. */
-
-#include "paddle/pten/kernels/empty_kernel.h"
-
-#include "paddle/pten/backends/cpu/cpu_context.h"
-#include "paddle/pten/core/kernel_registry.h"
-#include "paddle/pten/kernels/impl/empty_kernel_impl.h"
-
-PT_REGISTER_CTX_KERNEL(empty,
-                       CPU,
-                       ALL_LAYOUT,
-                       pten::Empty,
-                       bool,
-                       int,
-                       int64_t,
-                       float,
-                       double,
-                       paddle::platform::float16) {}
-
-PT_REGISTER_CTX_KERNEL(empty_like,
-                       CPU,
-                       ALL_LAYOUT,
-                       pten::EmptyLike,
-                       bool,
-                       int,
-                       int64_t,
-                       float,
-                       double,
-                       paddle::platform::float16) {}
