--- paddle/pten/kernels/impl/empty_kernel_impl.h
@@ -1,34 +1,0 @@
-/* Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. */
-
-#pragma once
-
-#include "paddle/pten/common/scalar_array.h"
-#include "paddle/pten/core/dense_tensor.h"
-
-namespace pten {
-
-template <typename T, typename ContextT>
-void Empty(const ContextT& dev_ctx,
-           const ScalarArray& shape,
-           DenseTensor* out) {
-  out->Resize(paddle::framework::make_ddim(shape.GetData()));
-}
-
-template <typename T, typename ContextT>
-void EmptyLike(const ContextT& dev_ctx, DenseTensor* out) {
-  out->mutable_data<T>();
-}
-
-}  // namespace pten
